abstract header creation helper
