+++ src/js/helpers/exportHelpers.js
@@ -5,6 +5,9 @@ import { ipcRenderer } from 'electron';
 //consts
 const OSX_DOCUMENTS_PATH = path.join(ospath.home(), 'Documents');
 const WIN_DOCUMENTS_PATH = path.join(ospath.home(), 'My Documents');
+//helpers
+import * as manifestHelpers from './manifestHelpers';
+import * as bibleHelpers from './bibleHelpers';
 
 /**
  * Prompts the user to enter a location/name to save the usfm project.
@@ -30,4 +33,36 @@ export function getFilePath(projectName, lastSaveLocation, ext) {
     defaultPath = path.join(ospath.home(), projectName + `.${ext}`);
   }
   return ipcRenderer.sendSync('save-as', { options: { defaultPath: defaultPath, filters: [{ extensions: [ext] }], title: 'Save Export As' } });
+}
+
+/**
+ * This function uses the manifest to populate the usfm JSON object id key in preparation
+ * of usfm to JSON conversion
+ * @param {string} projectSaveLocation - Path location in the filesystem for the project.
+ */
+export function getHeaderTags(projectSaveLocation) {
+  const manifest = manifestHelpers.getProjectManifest(projectSaveLocation);
+  const bookName = manifest.project.id;
+  /**Has fields such as "language_id": "en" and "resource_id": "ulb" and "direction":"ltr"*/
+  let sourceTranslation = manifest.source_translations[0];
+  let resourceName = sourceTranslation && sourceTranslation.language_id && sourceTranslation.resource_id ?
+    `${sourceTranslation.language_id.toUpperCase()}_${sourceTranslation.resource_id.toUpperCase()}` :
+    'N/A';
+  /**This will look like: ar_العربية_rtl to be included in the usfm id.
+   * This will make it easier to read for tC later on */
+  let targetLanguageCode = manifest.target_language ?
+    `${manifest.target_language.id}_${manifest.target_language.name}_${manifest.target_language.direction}` :
+    'N/A';
+  /**Date object when project was las changed in FS */
+  let lastEdited = fs.statSync(path.join(projectSaveLocation), bookName).atime;
+  let bookNameUppercase = bookName.toUpperCase();
+  /**Note the indication here of tc on the end of the id. This will act as a flag to ensure the correct parsing*/
+  return [{
+    "content": `${bookNameUppercase} ${resourceName} ${targetLanguageCode} ${lastEdited} tc`,
+    "tag": "id"
+  },
+  {
+    "content": bibleHelpers.convertToFullBookName(bookName),
+    "tag": "h"
+  }];
 }
